Chore: Create ModalTypes file
